requirements.txt
PyQt6>=6.5.0
watchdog>=3.0.0
transformers>=4.35.0
torch>=2.1.0
Pillow>=10.0.0
PyMuPDF>=1.23.0
PyInstaller>=6.5.0
